Added default board, functionality and spec file
I have hard-coded in a default playing board since the primary purpose
of this program is to investigate and understand unit testing and Rspec
output. The Adjudicator class now contains a play_game method that
allows for a user to play the game in the terminal. I have also added
some unit tests in battleships_spec.rb that focuses on testing some
methods in the Board class.